refactor: MVVM 패턴 구현 파일을 패키지로 분류
