+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
         android:theme="@style/Theme.Map"
         tools:targetApi="31">
         <activity
-            android:name=".LocationViewModel"
+            android:name=".ViewModel.LocationViewModel"
             android:exported="false" />
         <activity
-            android:name=".ViewActivity"
+            android:name=".View.ViewActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
