+++ app/src/main/java/campus/tech/kakao/map/Model/Location.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.Model
 
 data class Location (
     val name: String,
