+++ app/src/main/res/layout/activity_location_view_model.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LocationViewModel">
+    tools:context=".ViewModel.LocationViewModel">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
